Use a separate cache life test page
